rename dropdown folders
