+++ packages/renderer/src/renderer/screens/OpenGov/Dropdowns/index.tsx
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 export * from './DropdownReferendum';
+export * from './DropdownReferendaFilter';
