+++ packages/renderer/src/renderer/screens/OpenGov/Referenda/index.tsx
@@ -8,7 +8,7 @@ import * as Tabs from '@radix-ui/react-tabs';
 import * as Wrappers from './Wrappers';
 import { LinksFooter } from '@app/Utils';
 import { HistoryRow } from './HistoryRow';
-import { DropdownReferendaFilter } from '../DropdownMenus/DropdownReferendaFilter';
+import { DropdownReferendaFilter } from '../Dropdowns';
 
 import {
   ControlsWrapper,
